refactor: extract InboxSidebar from DesktopLayout
Extract the duplicated inbox tab content from both the tablet side panel
and the desktop side panel into a single InboxSidebar component. Uses a
variant prop ("tablet" | "desktop") with separate render paths since the
two variants differ significantly (desktop has drag-drop and inline
editing; tablet has swipe actions and touch handlers).

DesktopLayout drops from ~2905 lines to ~2370 lines.
